Corrected : issue with providing a wrong hash for input of size between 4GB+1 et 4GB+15 bytes Corrected : macro to define hardware 32-bits swap with GCC Thanks to Adrien Grand for both issues raised.
git-svn-id: https://xxhash.googlecode.com/svn/trunk@6 a90a0800-428d-89a3-ac20-94765a7798ee

+++ xxhash.c
@@ -98,7 +98,7 @@
 
 #if defined(_MSC_VER)     // Visual Studio
 #  define XXH_swap32 _byteswap_ulong
-#elif GCC_VERSION >= 430
+#elif GCC_VERSION >= 403
 #  define XXH_swap32 __builtin_bswap32
 #else
 static inline unsigned int XXH_swap32 (unsigned int x) {
@@ -208,7 +208,7 @@ struct XXH_state32_t
 	unsigned int v2;
 	unsigned int v3;
 	unsigned int v4;
-	unsigned int total_len;
+	unsigned long long total_len;
 	char memory[16];
 	int memsize;
 };
